extracted arguments into separate classes
