+++ plugin/src/main/java/net/thenextlvl/commander/paper/command/PermissionQueryCommand.java
@@ -1,0 +1,45 @@
+package net.thenextlvl.commander.paper.command;
+
+import com.mojang.brigadier.arguments.StringArgumentType;
+import com.mojang.brigadier.builder.ArgumentBuilder;
+import com.mojang.brigadier.context.CommandContext;
+import io.papermc.paper.command.brigadier.CommandSourceStack;
+import io.papermc.paper.command.brigadier.Commands;
+import lombok.RequiredArgsConstructor;
+import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
+import net.thenextlvl.commander.paper.CommanderPlugin;
+import org.bukkit.Bukkit;
+import org.bukkit.command.Command;
+
+@RequiredArgsConstructor
+@SuppressWarnings("UnstableApiUsage")
+class PermissionQueryCommand {
+    private final CommanderPlugin plugin;
+
+    public ArgumentBuilder<CommandSourceStack, ?> create() {
+        return Commands.literal("query")
+                .then(Commands.argument("command", StringArgumentType.string())
+                        .suggests((context, suggestions) -> {
+                            Bukkit.getCommandMap().getKnownCommands().values().stream()
+                                    .map(Command::getLabel)
+                                    .filter(s -> s.contains(suggestions.getRemaining()))
+                                    .map(StringArgumentType::escapeIfRequired)
+                                    .forEach(suggestions::suggest);
+                            return suggestions.buildFuture();
+                        })
+                        .executes(this::query));
+    }
+
+    private int query(CommandContext<CommandSourceStack> context) {
+        var sender = context.getSource().getSender();
+        var command = context.getArgument("command", String.class);
+        var registered = Bukkit.getCommandMap().getKnownCommands().get(command);
+        var permission = registered != null ? registered.getPermission() : null;
+        var message = registered == null ? "command.unknown" : permission != null ?
+                "permission.query.defined" : "permission.query.undefined";
+        plugin.bundle().sendMessage(sender, message,
+                Placeholder.parsed("permission", String.valueOf(permission)),
+                Placeholder.parsed("command", command));
+        return com.mojang.brigadier.Command.SINGLE_SUCCESS;
+    }
+}
