+++ plugin/src/main/java/net/thenextlvl/commander/paper/command/PermissionSetCommand.java
@@ -1,0 +1,56 @@
+package net.thenextlvl.commander.paper.command;
+
+import com.mojang.brigadier.arguments.StringArgumentType;
+import com.mojang.brigadier.builder.ArgumentBuilder;
+import com.mojang.brigadier.context.CommandContext;
+import io.papermc.paper.command.brigadier.CommandSourceStack;
+import io.papermc.paper.command.brigadier.Commands;
+import lombok.RequiredArgsConstructor;
+import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
+import net.thenextlvl.commander.paper.CommanderPlugin;
+import org.bukkit.Bukkit;
+import org.bukkit.command.Command;
+import org.bukkit.entity.Player;
+import org.bukkit.permissions.Permission;
+
+@RequiredArgsConstructor
+@SuppressWarnings("UnstableApiUsage")
+class PermissionSetCommand {
+    private final CommanderPlugin plugin;
+
+    public ArgumentBuilder<CommandSourceStack, ?> create() {
+        return Commands.literal("set")
+                .then(Commands.argument("command", StringArgumentType.string())
+                        .suggests((context, suggestions) -> {
+                            Bukkit.getCommandMap().getKnownCommands().values().stream()
+                                    .map(Command::getLabel)
+                                    .filter(s -> s.contains(suggestions.getRemaining()))
+                                    .map(StringArgumentType::escapeIfRequired)
+                                    .forEach(suggestions::suggest);
+                            return suggestions.buildFuture();
+                        })
+                        .then(Commands.argument("permission", StringArgumentType.string())
+                                .suggests((context, suggestions) -> {
+                                    Bukkit.getPluginManager().getPermissions().stream()
+                                            .map(Permission::getName)
+                                            .filter(s -> s.contains(suggestions.getRemaining()))
+                                            .map(StringArgumentType::escapeIfRequired)
+                                            .forEach(suggestions::suggest);
+                                    return suggestions.buildFuture();
+                                })
+                                .executes(this::set)));
+    }
+
+    private int set(CommandContext<CommandSourceStack> context) {
+        var sender = context.getSource().getSender();
+        var command = context.getArgument("command", String.class);
+        var permission = context.getArgument("permission", String.class);
+        var success = plugin.permissionOverride().override(command, permission);
+        var message = success ? "permission.set" : "nothing.changed";
+        plugin.bundle().sendMessage(sender, message,
+                Placeholder.parsed("permission", permission),
+                Placeholder.parsed("command", command));
+        if (success) Bukkit.getOnlinePlayers().forEach(Player::updateCommands);
+        return com.mojang.brigadier.Command.SINGLE_SUCCESS;
+    }
+}
